refactor(eval-label): name find-or-append once; share the eval-run test fixture
Two pieces of duplication found by reading the diff rather than by review.

`openCorpusLog` and `openOccurrenceLog` each carried their own find-or-append
loop, and each explained the same reason in its own long comment: a row holding
a field outside its identity, a capture time or an observation time, cannot use
`appendExact`, because rebuilding it tomorrow yields the same identity with a
different timestamp. That reason belongs to the log layer, not to two callers.
`OpenedJsonl.findOrAppend` now states it once, and both wrappers shrink to the
id rule and the row shape.

Two test files were each building a directory shaped like `agency eval run`
output, one a reduced copy of the other. That layout belongs to `readEvalRun`,
so a second copy is a second thing to update when the real one changes, and the
copy that gets missed keeps passing against a shape that no longer exists.
Both now use `writeRunFixture`.
