+++ packages/agency-lang/lib/eval/label/load/run.test.ts
@@ -4,6 +4,8 @@ import * as path from "path";
 
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
+import { writeRunFixture, type FixtureInput } from "../runFixture.js";
+
 import { describeIngestSkip } from "./eligibility.js";
 import { loadRun, selectLabelingFinalOutput, type LoadRunArgs } from "./run.js";
 import { DEFAULT_MAX_INGEST_BYTES } from "./types.js";
@@ -12,58 +14,8 @@ let root: string;
 let sourceDir: string;
 const warnings: string[] = [];
 
-type SourceInput = {
-  inputId: string;
-  status?: "success" | "error";
-  task?: unknown;
-  traceId?: string;
-  outputs?: unknown[];
-  omitRecord?: boolean;
-  legacyShape?: boolean;
-};
-
-function writeSource(inputs: SourceInput[], options: { dir?: string } = {}): string {
-  const dir = options.dir ?? sourceDir;
-  fs.mkdirSync(dir, { recursive: true });
-  fs.writeFileSync(path.join(dir, "config.json"), JSON.stringify({
-    provenance: { agent: { kind: "file", entry: "news.agency" } },
-  }));
-  fs.writeFileSync(path.join(dir, "summary.json"), JSON.stringify({
-    runId: path.basename(dir),
-    runDir: dir,
-    agentLabel: "news.agency:main",
-    okCount: inputs.length,
-    errorCount: 0,
-    inputs: inputs.map((input) => ({
-      inputId: input.inputId,
-      status: input.status ?? "success",
-      evalRecordPath: path.join(dir, "inputs", input.inputId, "agent", "eval-record.json"),
-      statelogPath: "",
-      workdirPath: "",
-    })),
-  }));
-  for (const input of inputs) {
-    const inputDir = path.join(dir, "inputs", input.inputId);
-    fs.mkdirSync(path.join(inputDir, "agent"), { recursive: true });
-    fs.writeFileSync(path.join(inputDir, "input.json"), JSON.stringify({
-      id: input.inputId,
-      task: input.task === undefined ? "do a thing" : input.task,
-    }));
-    if (input.omitRecord === true) {
-      continue;
-    }
-    const record = input.legacyShape === true
-      ? { traceId: input.traceId ?? "trace-1", finalResponse: "legacy" }
-      : {
-          traceId: input.traceId ?? "trace-1",
-          startedAtMs: 1000,
-          durationMs: 5,
-          evalOutputs: input.outputs ?? [{ value: "hello", threadId: "0", tMs: 1 }],
-          metrics: { models: ["gpt-4o"] },
-        };
-    fs.writeFileSync(path.join(inputDir, "agent", "eval-record.json"), JSON.stringify(record));
-  }
-  return dir;
+function writeSource(inputs: FixtureInput[], options: { dir?: string } = {}): string {
+  return writeRunFixture({ dir: options.dir ?? sourceDir, inputs });
 }
 
 function load(over: Partial<LoadRunArgs> = {}) {
